new commit

+++ 0x05-pointers_arrays_strings/5-rev_string.c
@@ -17,7 +17,7 @@ void rev_string(char *s)
 		revstr[b] = s[a];
 		b++;
 	}
-	for (c = 0; c <= strlen(revstr); c++)
+	for (c = 0; c <= (int)(strlen(revstr)); c++)
 	{
 		*s[c] = revstr[c];
 	}
